+++ test/common/http/conn_manager_impl_test.cc
@@ -2891,36 +2891,10 @@ TEST_F(HttpConnectionManagerImplTest, FrameFloodError) {
   Buffer::OwnedImpl fake_input("1234");
   EXPECT_LOG_NOT_CONTAINS("warning", "downstream HTTP flood",
                           conn_manager_->onData(fake_input, false));
-}
-
-// Verify that FrameFloodException causes connection to be closed abortively as well as logged
-// if runtime indicates to do so.
-TEST_F(HttpConnectionManagerImplTest, FrameFloodErrorWithLog) {
-  InSequence s;
-  setup(false, "");
-
-  EXPECT_CALL(*codec_, dispatch(_)).WillOnce(Invoke([&](Buffer::Instance&) -> void {
-    conn_manager_->newStream(response_encoder_);
-    throw FrameFloodException("too many outbound frames.");
-  }));
-
-  EXPECT_CALL(runtime_.snapshot_,
-              featureEnabled("http.connection_manager.log_flood_exception",
-                             Matcher<const envoy::type::v3::FractionalPercent&>(_)))
-      .WillOnce(Return(true));
-
-  EXPECT_CALL(response_encoder_.stream_, removeCallbacks(_));
-  EXPECT_CALL(filter_factory_, createFilterChain(_)).Times(0);
-
-  // FrameFloodException should result in reset of the streams followed by abortive close.
-  EXPECT_CALL(filter_callbacks_.connection_,
-              close(Network::ConnectionCloseType::FlushWriteAndDelay));
-
-  // Kick off the incoming data.
-  Buffer::OwnedImpl fake_input("1234");
-  EXPECT_LOG_CONTAINS("warning",
-                      "downstream HTTP flood from IP '0.0.0.0:0': too many outbound frames.",
-                      conn_manager_->onData(fake_input, false));
+  EXPECT_TRUE(filter_callbacks_.connection_.streamInfo().hasResponseFlag(
+      StreamInfo::ResponseFlag::DownstreamProtocolError));
+  EXPECT_EQ("codec error: too many outbound frames.",
+            filter_callbacks_.connection_.streamInfo().responseCodeDetails().value());
 }
 
 TEST_F(HttpConnectionManagerImplTest, IdleTimeoutNoCodec) {